Stop movie pausing on drag
This was a tough one (not to mention messy!). Basically because videojs's fixEvent mutilates the event object and forces `defaultPrevented` to false, we have to find a hack around it by storing the original event on the DOM element that raised it. We then use this to check if the event is prevented or not. Then clean up and delete the event off the DOM element. We also override the default video click handler in `player.js` to use mouseup instead of mousedown events. Fixes #876