nedb & Q compatibility
looks like we got some issue with backward compatibility.
Q revoked his 2.X release and started to works back on 1.X
and seem to not been supported by nedb.

No choice to use 2.0.2 for now till it’ll be better.

I ported the http api to use a correct callback as well