Remove WIN_PATHS logic
WIN_PATHS logic was previously added
(https://github.com/butterproject/butter-desktop/commit/2efc5c2664bd3e3cb9a2ce53d4944f2ed567d708)
to allow cross-platform compilation of this script, the downside is that
we needed to include all input paths twice, one for Windows and another
for Unix-like systems. Since the Unix/Linux version of makensis (v3)
supports both back slashes and forward slashes in input paths, it is
moot to keep both; with a single exception which is a specific case
where a Windows-only workaround for long paths does not work in
Unix/Linux environments and as such it is the only exception to be
retained.